Enum rat_salsa::ControlUI

pub enum ControlUI<Action = (), Err = ()> {
    Continue,
    Err(Err),
    NoChange,
    Change,
    Run(Action),
    Spawn(Action),
    Break,
}

UI control flow.

This is the result type for an event-handler.

Continue - Continue with execution.

Err(Err) - Equivalent to Result::Err. Use the macro crate::tr to convert from Result.

Unchanged - Event processed, no UI update necessary.

Changed - Event processed, UI update necessary.

Action(Action) - Run some computation on the model.

Spawn(Action) - Spawn some computation on the worker thread(s).

Break - Break the event loop; end the program.

There are multiple continuation functions that work with these states. And the macros crate::tr! and crate::check_break!

Variants

Continue

Continue execution.

Err(Err)

Error

NoChange

Event processed: no changes, no ui update.

Change

Event processed: changes happened, ui update.

Run(Action)

Run some action.

Spawn(Action)

Start some background action.

Break

Break the event loop.

Implementations

impl<Action, Err> ControlUI<Action, Err>

pub fn is_continue(&self) -> bool

Continue case

pub fn is_err(&self) -> bool

Err case

pub fn is_unchanged(&self) -> bool

Unchanged case

pub fn is_changed(&self) -> bool

Changed case

pub fn is_run(&self) -> bool

Action

pub fn is_spawn(&self) -> bool

Background action

pub fn is_break(&self) -> bool

Break case.

pub fn or(self, c: impl Into<ControlUI<Action, Err>>) -> ControlUI<Action, Err>

If the value is Continue, change to c.

pub fn or_else( self, f: impl FnOnce() -> ControlUI<Action, Err> ) -> ControlUI<Action, Err>

Run the continuation if the value is Continue.

pub fn or_do<R: Default>(&self, f: impl FnOnce() -> R) -> R

Run the continuation if the value is Continue. May return some R.

pub fn map_err<OtherErr>( self, f: impl FnOnce(Err) -> OtherErr ) -> ControlUI<Action, OtherErr>

Does some error conversion.

pub fn and_then<Other>( self, f: impl FnOnce(Action) -> ControlUI<Other, Err> ) -> ControlUI<Other, Err>

Run the continuation if the value is Run or Spawn.

Allows the result action to differ from the input to convert widget actions to more global ones.

Caveat: Allows no differentiation between Run and Spawn.

pub fn and_do<R: Default>(&self, f: impl FnOnce(&Action) -> R) -> R

Run the continuation if the value is Run or Spawn. May return some R.

Caveat: Allows no differentiation between Run and Spawn.

pub fn on_no_change( self, f: impl FnOnce() -> ControlUI<Action, Err> ) -> ControlUI<Action, Err>

Run the continuation if the value is Unchanged.

pub fn on_no_change_do<R: Default>(&self, f: impl FnOnce() -> R) -> R

Run the if the value is Unchanged. Allows to return some value.

pub fn on_change( self, f: impl FnOnce() -> ControlUI<Action, Err> ) -> ControlUI<Action, Err>

Run the continuation if the value is Changed.

pub fn on_change_do<R: Default>(&self, f: impl FnOnce() -> R) -> R

Run if the value is Changed. Allows to return some value.