Enum rat_salsa::Locale

pub enum Locale {

    POSIX,
    aa_DJ,
    aa_ER,
    aa_ER_saaho,
    aa_ET,
    af_ZA,
    agr_PE,
    ak_GH,
    am_ET,
    an_ES,
    anp_IN,
    ar_AE,
    ar_BH,
    ar_DZ,
    ar_EG,
    ar_IN,
    ar_IQ,
    ar_JO,
    ar_KW,
    ar_LB,
    ar_LY,
    ar_MA,
    ar_OM,
    ar_QA,
    ar_SA,
    ar_SD,
    ar_SS,
    ar_SY,
    ar_TN,
    ar_YE,
    as_IN,
    ast_ES,
    ayc_PE,
    az_AZ,
    az_IR,
    be_BY,
    be_BY_latin,
    bem_ZM,
    ber_DZ,
    ber_MA,
    bg_BG,
    bhb_IN,
    bho_IN,
    bho_NP,
    bi_VU,
    bn_BD,
    bn_IN,
    bo_CN,
    bo_IN,
    br_FR,
    br_FR_euro,
    brx_IN,
    bs_BA,
    byn_ER,
    ca_AD,
    ca_ES,
    ca_ES_euro,
    ca_ES_valencia,
    ca_FR,
    ca_IT,
    ce_RU,
    chr_US,
    cmn_TW,
    crh_UA,
    cs_CZ,
    csb_PL,
    cv_RU,
    cy_GB,
    da_DK,
    de_AT,
    de_AT_euro,
    de_BE,
    de_BE_euro,
    de_CH,
    de_DE,
    de_DE_euro,
    de_IT,
    de_LI,
    de_LU,
    de_LU_euro,
    doi_IN,
    dsb_DE,
    dv_MV,
    dz_BT,
    el_CY,
    el_GR,
    el_GR_euro,
    en_AG,
    en_AU,
    en_BW,
    en_CA,
    en_DK,
    en_GB,
    en_HK,
    en_IE,
    en_IE_euro,
    en_IL,
    en_IN,
    en_NG,
    en_NZ,
    en_PH,
    en_SC,
    en_SG,
    en_US,
    en_ZA,
    en_ZM,
    en_ZW,
    eo,
    es_AR,
    es_BO,
    es_CL,
    es_CO,
    es_CR,
    es_CU,
    es_DO,
    es_EC,
    es_ES,
    es_ES_euro,
    es_GT,
    es_HN,
    es_MX,
    es_NI,
    es_PA,
    es_PE,
    es_PR,
    es_PY,
    es_SV,
    es_US,
    es_UY,
    es_VE,
    et_EE,
    eu_ES,
    eu_ES_euro,
    fa_IR,
    ff_SN,
    fi_FI,
    fi_FI_euro,
    fil_PH,
    fo_FO,
    fr_BE,
    fr_BE_euro,
    fr_CA,
    fr_CH,
    fr_FR,
    fr_FR_euro,
    fr_LU,
    fr_LU_euro,
    fur_IT,
    fy_DE,
    fy_NL,
    ga_IE,
    ga_IE_euro,
    gd_GB,
    gez_ER,
    gez_ER_abegede,
    gez_ET,
    gez_ET_abegede,
    gl_ES,
    gl_ES_euro,
    gu_IN,
    gv_GB,
    ha_NG,
    hak_TW,
    he_IL,
    hi_IN,
    hif_FJ,
    hne_IN,
    hr_HR,
    hsb_DE,
    ht_HT,
    hu_HU,
    hy_AM,
    ia_FR,
    id_ID,
    ig_NG,
    ik_CA,
    is_IS,
    it_CH,
    it_IT,
    it_IT_euro,
    iu_CA,
    ja_JP,
    ka_GE,
    kab_DZ,
    kk_KZ,
    kl_GL,
    km_KH,
    kn_IN,
    ko_KR,
    kok_IN,
    ks_IN,
    ks_IN_devanagari,
    ku_TR,
    kw_GB,
    ky_KG,
    lb_LU,
    lg_UG,
    li_BE,
    li_NL,
    lij_IT,
    ln_CD,
    lo_LA,
    lt_LT,
    lv_LV,
    lzh_TW,
    mag_IN,
    mai_IN,
    mai_NP,
    mfe_MU,
    mg_MG,
    mhr_RU,
    mi_NZ,
    miq_NI,
    mjw_IN,
    mk_MK,
    ml_IN,
    mn_MN,
    mni_IN,
    mnw_MM,
    mr_IN,
    ms_MY,
    mt_MT,
    my_MM,
    nan_TW,
    nan_TW_latin,
    nb_NO,
    nds_DE,
    nds_NL,
    ne_NP,
    nhn_MX,
    niu_NU,
    niu_NZ,
    nl_AW,
    nl_BE,
    nl_BE_euro,
    nl_NL,
    nl_NL_euro,
    nn_NO,
    nr_ZA,
    nso_ZA,
    oc_FR,
    om_ET,
    om_KE,
    or_IN,
    os_RU,
    pa_IN,
    pa_PK,
    pap_AW,
    pap_CW,
    pl_PL,
    ps_AF,
    pt_BR,
    pt_PT,
    pt_PT_euro,
    quz_PE,
    raj_IN,
    ro_RO,
    ru_RU,
    ru_UA,
    rw_RW,
    sa_IN,
    sah_RU,
    sat_IN,
    sc_IT,
    sd_IN,
    sd_IN_devanagari,
    se_NO,
    sgs_LT,
    shn_MM,
    shs_CA,
    si_LK,
    sid_ET,
    sk_SK,
    sl_SI,
    sm_WS,
    so_DJ,
    so_ET,
    so_KE,
    so_SO,
    sq_AL,
    sq_MK,
    sr_ME,
    sr_RS,
    sr_RS_latin,
    ss_ZA,
    st_ZA,
    sv_FI,
    sv_FI_euro,
    sv_SE,
    sw_KE,
    sw_TZ,
    szl_PL,
    ta_IN,
    ta_LK,
    tcy_IN,
    te_IN,
    tg_TJ,
    th_TH,
    the_NP,
    ti_ER,
    ti_ET,
    tig_ER,
    tk_TM,
    tl_PH,
    tn_ZA,
    to_TO,
    tpi_PG,
    tr_CY,
    tr_TR,
    ts_ZA,
    tt_RU,
    tt_RU_iqtelif,
    ug_CN,
    uk_UA,
    unm_US,
    ur_IN,
    ur_PK,
    uz_UZ,
    uz_UZ_cyrillic,
    ve_ZA,
    vi_VN,
    wa_BE,
    wa_BE_euro,
    wae_CH,
    wal_ET,
    wo_SN,
    xh_ZA,
    yi_US,
    yo_NG,
    yue_HK,
    yuw_PG,
    zh_CN,
    zh_HK,
    zh_SG,
    zh_TW,
    zu_ZA,

}

Locales matching the locales in glibc.

Most locale names follow the syntax language[_territory][@modifier]. The @ is replaced with _ in the enum variant names.

The default locale is POSIX.

License note: The Free Software Foundation does not claim any copyright interest in the locale data of the GNU C Library; they believe it is not copyrightable.

Variants

POSIX

POSIX: POSIX Standard Locale.

aa_DJ

aa_DJ: Afar language locale for Djibouti (Cadu/Laaqo Dialects).

aa_ER

aa_ER: Afar language locale for Eritrea (Cadu/Laaqo Dialects).

aa_ER_saaho

aa_ER@saaho: Afar language locale for Eritrea (Saaho Dialect).

aa_ET

aa_ET: Afar language locale for Ethiopia (Cadu/Carra Dialects).

af_ZA

af_ZA: Afrikaans locale for South Africa.

agr_PE

agr_PE: Awajún (agr) locale for Peru.

ak_GH

ak_GH: Akan locale for Ghana.

am_ET

am_ET: Amharic language locale for Ethiopia.

an_ES

an_ES: Aragonese locale for Spain.

anp_IN

anp_IN: Angika language locale for India.

ar_AE

ar_AE: Arabic language locale for United Arab Emirates.

ar_BH

ar_BH: Arabic language locale for Bahrain.

ar_DZ

ar_DZ: Arabic language locale for Algeria.

ar_EG

ar_EG: Arabic language locale for Egypt.

ar_IN

ar_IN: Arabic language locale for India.

ar_IQ

ar_IQ: Arabic language locale for Iraq.

ar_JO

ar_JO: Arabic language locale for Jordan.

ar_KW

ar_KW: Arabic language locale for Kuwait.

ar_LB

ar_LB: Arabic language locale for Lebanon.

ar_LY

ar_LY: Arabic language locale for Libyan Arab Jamahiriya.

ar_MA

ar_MA: Arabic language locale for Morocco.

ar_OM

ar_OM: Arabic language locale for Oman.

ar_QA

ar_QA: Arabic language locale for Qatar.

ar_SA

ar_SA: Arabic locale for Saudi Arabia.

ar_SD

ar_SD: Arabic language locale for Sudan.

ar_SS

ar_SS: Arabic language locale for South Sudan.

ar_SY

ar_SY: Arabic language locale for Syrian Arab Republic.

ar_TN

ar_TN: Arabic language locale for Tunisia.

ar_YE

ar_YE: Arabic language locale for Yemen.

as_IN

as_IN: Assamese language locale for India.

ast_ES

ast_ES: Asturian locale for Spain.

ayc_PE

ayc_PE: Aymara (ayc) locale for Peru.

az_AZ

az_AZ: Azeri language locale for Azerbaijan (latin).

az_IR

az_IR: South Azerbaijani language locale for Iran.

be_BY

be_BY: Belarusian locale for Belarus.

be_BY_latin

be_BY@latin: Belarusian Latin-Script locale for Belarus.

bem_ZM

bem_ZM: Bemba locale for Zambia.

ber_DZ

ber_DZ: Berber language locale for Algeria (latin).

ber_MA

ber_MA: Berber language locale for Morocco (tifinagh).

bg_BG

bg_BG: Bulgarian locale for Bulgaria.

bhb_IN

bhb_IN: Bhili(devanagari) language locale for India.

bho_IN

bho_IN: Bhojpuri language locale for India.

bho_NP

bho_NP: Bhojpuri language locale for Nepal.

bi_VU

bi_VU: Bislama language locale for Vanuatu.

bn_BD

bn_BD: Bangla language locale for Bangladesh.

bn_IN

bn_IN: Bangla language locale for India.

bo_CN

bo_CN: Tibetan language locale for P.R. of China.

bo_IN

bo_IN: Tibetan language locale for India.

br_FR

br_FR: Breton language locale for France.

br_FR_euro

br_FR@euro: Breton locale for France with Euro.

brx_IN

brx_IN: Bodo language locale for India.

bs_BA

bs_BA: Bosnian language locale for Bosnia and Herzegowina.

byn_ER

byn_ER: Blin language locale for Eritrea.

ca_AD

ca_AD: Catalan locale for Andorra .

ca_ES

ca_ES: Catalan locale for Spain.

ca_ES_euro

ca_ES@euro: Catalan locale for Catalonia with Euro.

ca_ES_valencia

ca_ES@valencia: Valencian (southern Catalan) locale for Spain with Euro.

ca_FR

ca_FR: Catalan locale for France .

ca_IT

ca_IT: Catalan locale for Italy (L’Alguer) .

ce_RU

ce_RU: Chechen locale for RUSSIAN FEDERATION.

chr_US

chr_US: Cherokee language locale for United States.

cmn_TW

cmn_TW: Mandarin Chinese locale for the Republic of China.

crh_UA

crh_UA: Crimean Tatar (Crimean Turkish) language locale for Ukraine.

cs_CZ

cs_CZ: Czech locale for the Czech Republic.

csb_PL

csb_PL: Kashubian locale for Poland.

cv_RU

cv_RU: Chuvash locale for Russia.

cy_GB

cy_GB: Welsh language locale for Great Britain.

da_DK

da_DK: Danish locale for Denmark.

de_AT

de_AT: German locale for Austria.

de_AT_euro

de_AT@euro: German locale for Austria with Euro.

de_BE

de_BE: German locale for Belgium.

de_BE_euro

de_BE@euro: German locale for Belgium with Euro.

de_CH

de_CH: German locale for Switzerland.

de_DE

de_DE: German locale for Germany.

de_DE_euro

de_DE@euro: German locale for Germany with Euro.

de_IT

de_IT: German language locale for Italy.

de_LI

de_LI: German locale for Liechtenstein.

de_LU

de_LU: German locale for Luxemburg.

de_LU_euro

de_LU@euro: German locale for Luxemburg with Euro.

doi_IN

doi_IN: Dogri language locale for India.

dsb_DE

dsb_DE: Lower Sorbian locale for Germany.

dv_MV

dv_MV: Dhivehi Language Locale for Maldives.

dz_BT

dz_BT: Dzongkha language locale for Bhutan.

el_CY

el_CY: Greek locale for Cyprus.

el_GR

el_GR: Greek locale for Greece.

el_GR_euro

el_GR@euro: Greek locale for Greece with Euro.

en_AG

en_AG: English language locale for Antigua and Barbuda.

en_AU

en_AU: English locale for Australia.

en_BW

en_BW: English locale for Botswana.

en_CA

en_CA: English locale for Canada.

en_DK

en_DK: English locale for Denmark.

en_GB

en_GB: English locale for Britain.

en_HK

en_HK: English locale for Hong Kong.

en_IE

en_IE: English locale for Ireland.

en_IE_euro

en_IE@euro: English locale for Ireland with Euro.

en_IL

en_IL: English locale for Israel.

en_IN

en_IN: English language locale for India.

en_NG

en_NG: English locale for Nigeria.

en_NZ

en_NZ: English locale for New Zealand.

en_PH

en_PH: English language locale for Philippines.

en_SC

en_SC: English locale for the Seychelles.

en_SG

en_SG: English language locale for Singapore.

en_US

en_US: English locale for the USA.

en_ZA

en_ZA: English locale for South Africa.

en_ZM

en_ZM: English locale for Zambia.

en_ZW

en_ZW: English locale for Zimbabwe.

eo

eo: Esperanto language locale.

es_AR

es_AR: Spanish locale for Argentina.

es_BO

es_BO: Spanish locale for Bolivia.

es_CL

es_CL: Spanish locale for Chile.

es_CO

es_CO: Spanish locale for Colombia.

es_CR

es_CR: Spanish locale for Costa Rica.

es_CU

es_CU: Spanish locale for Cuba.

es_DO

es_DO: Spanish locale for Dominican Republic.

es_EC

es_EC: Spanish locale for Ecuador.

es_ES

es_ES: Spanish locale for Spain.

es_ES_euro

es_ES@euro: Spanish locale for Spain with Euro.

es_GT

es_GT: Spanish locale for Guatemala.

es_HN

es_HN: Spanish locale for Honduras.

es_MX

es_MX: Spanish locale for Mexico.

es_NI

es_NI: Spanish locale for Nicaragua.

es_PA

es_PA: Spanish locale for Panama.

es_PE

es_PE: Spanish locale for Peru.

es_PR

es_PR: Spanish locale for Puerto Rico.

es_PY

es_PY: Spanish locale for Paraguay.

es_SV

es_SV: Spanish locale for El Salvador.

es_US

es_US: Spanish locale for the USA.

es_UY

es_UY: Spanish locale for Uruguay.

es_VE

es_VE: Spanish locale for Venezuela.

et_EE

et_EE: Estonian locale for Estonia.

eu_ES

eu_ES: Basque locale for Spain.

eu_ES_euro

eu_ES@euro: Basque language locale for Spain with Euro.

fa_IR

fa_IR: Persian locale for Iran.

ff_SN

ff_SN: Fulah locale for Senegal.

fi_FI

fi_FI: Finnish locale for Finland.

fi_FI_euro

fi_FI@euro: Finnish locale for Finland with Euro.

fil_PH

fil_PH: Filipino language locale for Philippines.

fo_FO

fo_FO: Faroese locale for Faroe Islands.

fr_BE

fr_BE: French locale for Belgium.

fr_BE_euro

fr_BE@euro: French locale for Belgium with Euro.

fr_CA

fr_CA: French locale for Canada.

fr_CH

fr_CH: French locale for Switzerland.

fr_FR

fr_FR: French locale for France.

fr_FR_euro

fr_FR@euro: French locale for France with Euro.

fr_LU

fr_LU: French locale for Luxemburg.

fr_LU_euro

fr_LU@euro: French locale for Luxemburg with Euro.

fur_IT

fur_IT: Furlan locale for Italy.

fy_DE

fy_DE: Sater Frisian and North Frisian Locale for Germany.

fy_NL

fy_NL: Frisian locale for the Netherlands.

ga_IE

ga_IE: Irish locale for Ireland.

ga_IE_euro

ga_IE@euro: Irish locale for Ireland with Euro.

gd_GB

gd_GB: Scots Gaelic language locale for Great Britain.

gez_ER

gez_ER: Ge’ez language locale for Eritrea.

gez_ER_abegede

gez_ER@abegede: Ge’ez language locale for Eritrea With Abegede Collation.

gez_ET

gez_ET: Ge’ez language locale for Ethiopia.

gez_ET_abegede

gez_ET@abegede: Ge’ez language locale for Ethiopia With Abegede Collation.

gl_ES

gl_ES: Galician locale for Spain.

gl_ES_euro

gl_ES@euro: Galician locale for Spain with Euro.

gu_IN

gu_IN: Gujarati Language Locale For India.

gv_GB

gv_GB: Manx Gaelic locale for Britain.

ha_NG

ha_NG: Hausa locale for Nigeria.

hak_TW

hak_TW: Hakka Chinese locale for the Republic of China.

he_IL

he_IL: Hebrew locale for Israel.

hi_IN

hi_IN: Hindi language locale for India.

hif_FJ

hif_FJ: Fiji Hindi (Latin) language locale for Fiji.

hne_IN

hne_IN: Chhattisgarhi language locale for India.

hr_HR

hr_HR: Croatian locale for Croatia.

hsb_DE

hsb_DE: Upper Sorbian locale for Germany.

ht_HT

ht_HT: Kreyol locale for Haiti.

hu_HU

hu_HU: Hungarian locale for Hungary.

hy_AM

hy_AM: Armenian language locale for Armenia.

ia_FR

ia_FR: Interlingua locale for France.

id_ID

id_ID: Indonesian locale for Indonesia.

ig_NG

ig_NG: Igbo locale for Nigeria.

ik_CA

ik_CA: Inupiaq locale for Canada.

is_IS

is_IS: Icelandic locale for Iceland.

it_CH

it_CH: Italian locale for Switzerland.

it_IT

it_IT: Italian locale for Italy.

it_IT_euro

it_IT@euro: Italian locale for Italy with Euro.

iu_CA

iu_CA: Inuktitut language locale for Nunavut, Canada.

ja_JP

ja_JP: Japanese language locale for Japan.

ka_GE

ka_GE: Georgian language locale for Georgia.

kab_DZ

kab_DZ: Kabyle language locale for Algeria.

kk_KZ

kk_KZ: Kazakh locale for Kazakhstan.

kl_GL

kl_GL: Greenlandic locale for Greenland.

km_KH

km_KH: Khmer locale for Cambodia.

kn_IN

kn_IN: Kannada language locale for India.

ko_KR

ko_KR: Korean locale for Republic of Korea.

kok_IN

kok_IN: Konkani language locale for India.

ks_IN

ks_IN: Kashmiri language locale for India.

ks_IN_devanagari

ks_IN@devanagari: Kashmiri(devanagari) language locale for India.

ku_TR

ku_TR: Kurdish (latin) locale for Turkey.

kw_GB

kw_GB: Cornish locale for Britain.

ky_KG

ky_KG: Kyrgyz Language Locale for Kyrgyzstan.

lb_LU

lb_LU: Luxembourgish locale for Luxembourg.

lg_UG

lg_UG: Luganda locale for Uganda.

li_BE

li_BE: Limburgish Language Locale for Belgium.

li_NL

li_NL: Limburgish Language Locale for the Netherlands.

lij_IT

lij_IT: Ligurian locale for Italy.

ln_CD

ln_CD: Lingala locale for Democratic Republic of the Congo.

lo_LA

lo_LA: Lao locale for Laos.

lt_LT

lt_LT: Lithuanian locale for Lithuania.

lv_LV

lv_LV: Latvian locale for Latvia.

lzh_TW

lzh_TW: Literary Chinese locale for the Republic of China.

mag_IN

mag_IN: Magahi language locale for India.

mai_IN

mai_IN: Maithili language locale for India.

mai_NP

mai_NP: Maithili language locale for Nepal.

mfe_MU

mfe_MU: Morisyen locale for Mauritius.

mg_MG

mg_MG: Malagasy locale for Madagascar.

mhr_RU

mhr_RU: Mari locale for Russia.

mi_NZ

mi_NZ: Maori language locale for New Zealand.

miq_NI

miq_NI: Miskito language locale for Nicaragua.

mjw_IN

mjw_IN: Karbi language locale for India.

mk_MK

mk_MK: Macedonian locale for Macedonia.

ml_IN

ml_IN: Malayalam language locale for India.

mn_MN

mn_MN: Mongolian locale for Mongolia.

mni_IN

mni_IN: Manipuri language locale for India.

mnw_MM

mnw_MM: Mon language locale for Myanmar.

mr_IN

mr_IN: Marathi language locale for India.

ms_MY

ms_MY: Malay language locale for Malaysia.

mt_MT

mt_MT: Maltese language locale for Malta.

my_MM

my_MM: Burmese language locale for Myanmar.

nan_TW

nan_TW: Min Nan Chinese locale for the Republic of China.

nan_TW_latin

nan_TW@latin: Minnan language locale for Taiwan.

nb_NO

nb_NO: Norwegian (Bokmål) locale for Norway.

nds_DE

nds_DE: Low(lands) Saxon Language Locale for Germany.

nds_NL

nds_NL: Low(lands) Saxon Language Locale for the Netherlands.

ne_NP

ne_NP: Nepali language locale for Nepal.

nhn_MX

nhn_MX: Central Nahuatl for Mexico.

niu_NU

niu_NU: Niuean (Vagahau Niue) locale for Niue.

niu_NZ

niu_NZ: Niuean (Vagahau Niue) locale for New Zealand.

nl_AW

nl_AW: Dutch language locale for Aruba.

nl_BE

nl_BE: Dutch locale for Belgium.

nl_BE_euro

nl_BE@euro: Dutch locale for Belgium with Euro.

nl_NL

nl_NL: Dutch locale for the Netherlands.

nl_NL_euro

nl_NL@euro: Dutch locale for the Netherlands with Euro.

nn_NO

nn_NO: Nynorsk language locale for Norway.

nr_ZA

nr_ZA: Southern Ndebele locale for South Africa.

nso_ZA

nso_ZA: Northern Sotho locale for South Africa.

oc_FR

oc_FR: Occitan Language Locale for France.

om_ET

om_ET: Oromo language locale for Ethiopia.

om_KE

om_KE: Oromo language locale for Kenya.

or_IN

or_IN: Odia language locale for India.

os_RU

os_RU: Ossetian locale for Russia.

pa_IN

pa_IN: Punjabi language locale for Indian Punjabi(Gurmukhi).

pa_PK

pa_PK: Punjabi (Shahmukhi) Language Locale for Pakistan.

pap_AW

pap_AW: Papiamento Language for Aruba.

pap_CW

pap_CW: Papiamento language for Curaçao.

pl_PL

pl_PL: Polish locale for Poland.

ps_AF

ps_AF: Pashto locale for Afghanistan.

pt_BR

pt_BR: Portuguese locale for Brasil.

pt_PT

pt_PT: Portuguese locale for Portugal.

pt_PT_euro

pt_PT@euro: Portuguese locale for Portugal with Euro.

quz_PE

quz_PE: Cusco Quechua locale for Peru.

raj_IN

raj_IN: Rajasthani language locale for India.

ro_RO

ro_RO: Romanian locale for Romania.

ru_RU

ru_RU: Russian locale for Russia.

ru_UA

ru_UA: Russian locale for Ukraine.

rw_RW

rw_RW: Kinyarwanda language locale for Rwanda.

sa_IN

sa_IN: Sanskrit language locale for India.

sah_RU

sah_RU: Sakha (Yakut) locale for Russian Federation.

sat_IN

sat_IN: Santali language locale for India.

sc_IT

sc_IT: Sardinian locale for Italy.

sd_IN

sd_IN: Sindhi language locale for India.

sd_IN_devanagari

sd_IN@devanagari: Sindhi language locale for India.

se_NO

se_NO: Northern Saami language locale for Norway.

sgs_LT

sgs_LT: Samogitian language locale for Lithuania.

shn_MM

shn_MM: Shan language locale for Myanmar.

shs_CA

shs_CA: Shuswap locale for Canada.

si_LK

si_LK: Sinhala language locale for Sri Lanka.

sid_ET

sid_ET: Sidama language locale for Ethiopia.

sk_SK

sk_SK: Slovak locale for Slovak.

sl_SI

sl_SI: Slovenian locale for Slovenia.

sm_WS

sm_WS: Samoan language locale for Samoa.

so_DJ

so_DJ: Somali language locale for Djibouti.

so_ET

so_ET: Somali language locale for Ethiopia.

so_KE

so_KE: Somali language locale for Kenya.

so_SO

so_SO: Somali language locale for Somalia.

sq_AL

sq_AL: Albanian language locale for Albania.

sq_MK

sq_MK: Albanian language locale for Macedonia.

sr_ME

sr_ME: Serbian locale for Montenegro.

sr_RS

sr_RS: Serbian locale for Serbia.

sr_RS_latin

sr_RS@latin: Serbian Latin locale for Serbia.

ss_ZA

ss_ZA: Swati locale for South Africa.

st_ZA

st_ZA: Sotho locale for South Africa.

sv_FI

sv_FI: Swedish locale for Finland.

sv_FI_euro

sv_FI@euro: Swedish locale for Finland with Euro.

sv_SE

sv_SE: Swedish locale for Sweden.

sw_KE

sw_KE: Swahili locale for Kenya.

sw_TZ

sw_TZ: Swahili locale for Tanzania.

szl_PL

szl_PL: Silesian locale for Poland.

ta_IN

ta_IN: Tamil language locale for India.

ta_LK

ta_LK: Tamil language locale for Sri Lanka.

tcy_IN

tcy_IN: Tulu language locale for India.

te_IN

te_IN: Telugu language locale for India.

tg_TJ

tg_TJ: Tajik language locale for Tajikistan.

th_TH

th_TH: Thai locale for Thailand.

the_NP

the_NP: Tharu language locale for Nepal.

ti_ER

ti_ER: Tigrigna language locale for Eritrea.

ti_ET

ti_ET: Tigrigna language locale for Ethiopia.

tig_ER

tig_ER: Tigre language locale for Eritrea.

tk_TM

tk_TM: Turkmen locale for Turkmenistan.

tl_PH

tl_PH: Tagalog language locale for Philippines.

tn_ZA

tn_ZA: Tswana locale for South Africa.

to_TO

to_TO: Tongan language locale for Tonga.

tpi_PG

tpi_PG: Tok Pisin language locale for Papua New Guinea.

tr_CY

tr_CY: Turkish language locale for Cyprus.

tr_TR

tr_TR: Turkish locale for Turkey.

ts_ZA

ts_ZA: Tsonga locale for South Africa.

tt_RU

tt_RU: Tatar language locale for Russia.

tt_RU_iqtelif

tt_RU@iqtelif: Tatar language locale using IQTElif alphabet; for Tatarstan, Russian Federation.

ug_CN

ug_CN: Uyghur locale for China.

uk_UA

uk_UA: Ukrainian Language Locale for Ukraine.

unm_US

unm_US: Unami Delaware locale for the USA.

ur_IN

ur_IN: Urdu language locale for India.

ur_PK

ur_PK: Urdu Language Locale for Pakistan.

uz_UZ

uz_UZ: Uzbek (latin) locale for Uzbekistan.

uz_UZ_cyrillic

uz_UZ@cyrillic: Uzbek (cyrillic) locale for Uzbekistan.

ve_ZA

ve_ZA: Venda locale for South Africa.

vi_VN

vi_VN: Vietnamese language locale for Vietnam.

wa_BE

wa_BE: Walloon Language Locale for Belgium.

wa_BE_euro

wa_BE@euro: Walloon locale for Belgium with Euro.

wae_CH

wae_CH: Walser locale for Switzerland.

wal_ET

wal_ET: Walaita language locale for Ethiopia.

wo_SN

wo_SN: Wolof locale for Senegal.

xh_ZA

xh_ZA: Xhosa locale for South Africa.

yi_US

yi_US: Yiddish Language locale for the USA.

yo_NG

yo_NG: Yoruba locale for Nigeria.

yue_HK

yue_HK: Yue Chinese (Cantonese) language locale for Hong Kong.

yuw_PG

yuw_PG: YauNungon locale for Papua New Guinea.

zh_CN

zh_CN: Chinese locale for Peoples Republic of China.

zh_HK

zh_HK: Chinese language locale for Hong Kong.

zh_SG

zh_SG: Chinese language locale for Singapore.

zh_TW

zh_TW: Chinese locale for Taiwan R.O.C.

zu_ZA

zu_ZA: Zulu locale for South Africa.