Module rat_salsa::number

Number formatting.

This one uses a pattern string instead of the format! style.

use std::rc::Rc;
use pure_rust_locales::Locale::de_AT_euro;
use rat_salsa::number;
use rat_salsa::number::{NumberFormat, NumberSymbols};

// formats accordingly, uses the default symbols.
let s = number::format(4561.2234, "###,##0.00").expect("works");
assert_eq!(s, "  4,561.22");

// uses symbols
let sym = NumberSymbols::monetary(de_AT_euro);
let s = number::format(4561.2234, "$ ###,##0.00").expect("works");
assert_eq!(s.as_str(), "€   4.561,22");

// prepared format
let sym = Rc::new(NumberSymbols::monetary(de_AT_euro));
let m2 = NumberFormat::news("$ ###,##0.00", &sym).expect("works");

let s = m2.fmt(4561.2234);

use crate::rat_salsa::number::FormatNumber;
println!("combined output: {}", 4561.2234f64.fmt(&m2));

The following patterns are recognized:

0 - digit or 0

9 - digit or space

# - digit or sign or space

- - sign; show space for positive

+ - sign; show ‘+’ for positive and ‘-’ for negative. not localized.

. - decimal separator

: - decimal separator, always shown

, - grouping separator. maybe left out if the symbols say so.

E - upper case exponent

e - lower case exponent

- space can be used as separator

‘$’ - currency. variable length output according to the currency-symbol.

\ - all ascii characters (ascii 32-128!) are reserved and must be escaped.

_ - other unicode characters can be used without escaping.

Modules

core

Structs

CurrencySym

Currency symbol. Const constructable short inline string.

FormattedNumber

Holds a temporary result from FormatNumber. The only purpose is as anchor for the Display trait.

NumberFormat

Holds the pattern for the number format and some additional data.

NumberSymbols

Symbols for number formatting.

RefFormattedNumber

Holds a temporary result from FormatNumber. The only purpose is as anchor for the Display trait.

Enums

NumberFmtError

Errors

Traits

FormatNumber

Format a number according to a format string.

ParseNumber

Parses a number from a &str.

Functions

fmt

Format a Number according to the format.

fmt_to

Format a Number according to the format.

format

Format a Number according to the format string. Uses the default symbols.

format_to

Format a Number according to the format string. Uses the default symbols.

formats

Format a Number according to the format string.

formats_to

Format a Number according to the format string.

parse_fmt

Parse using the NumberFormat. Parses the number after applying core::unmap_num

parse_format

Parse using the NumberFormat. Parses the number after applying core::unmap_num

parse_sym

Parse using the NumberSymbols. Parses the number after applying core::clean_num