Module rat_salsa::number::core

Functions

clean_num

Get the clean number.

format_to

Formats the number and writes the result to out.

map_num

Takes a raw number string and applies the format.

parse_fmt

Parse the number according to the exact format.

parse_sym

Parse the number only using the symbols for translation. Takes digits and some specials and ignores the rest.

unmap_num

Unmap the formatted string back to a format that f64::parse() can understand.