Struct rat_salsa::Focus

pub struct Focus<'a> {
    pub areas: Vec<Rect>,
    pub focus: Vec<&'a FocusFlag>,
}

Keeps track of the focus.

It works by adding a FocusFlag to the State of a widget. Focus is constructed with a list of references to these flags and switches the focus that way. Each widget stays separate otherwise and can pull its focus state from this struct.

Focus::new([
    (&widget1.focus, widget1.area),
    (&widget2.focus, widget2.area),
]).handle(evt, DefaultKeys)
.and_do(|_| uistate.repaint.set());

repaint in the example is a Repaint. This is necessary as the focus change does not automatically trigger a repaint.

Fields
areas: Vec<Rect>

Areas for each widget.

focus: Vec<&'a FocusFlag>

List of flags.

Implementations

impl<'a> Focus<'a>

pub fn new(np: impl IntoIterator<Item = (&'a FocusFlag, Rect)>) -> Self

Create a focus cycle.

Take a reference to a FocusFlag and a Rect for mouse-events.

pub fn append(self, np: impl IntoIterator<Item = (&'a FocusFlag, Rect)>) -> Self

Add more to the focus cycle.

pub fn focus_no_lost(&self, tag: u16)

Sets the focus to the widget with tag.

Sets focus and gained but not lost. This can be used to prevent validation of the field.

pub fn focus(&self, tag: u16)

Sets the focus to the widget with tag.

Sets the focus, gained and lost flags. If this ends up with the same widget as before focus, gained and lost flag are all set.

pub fn reset_lost_gained(&self)

Reset lost + gained flags.

This is done automatically in HandleCrossterm::handle() for every event. This means these flags are only ever set if handle() returns Run(true) to allow immediate reactions to focus changes.

pub fn focus_idx(&self, idx: usize)

Change the focus.

Sets the focus, gained and lost flags.

If the field at idx has the focus all three are set.

pub fn next(&self) -> bool

Focus the next widget in the cycle.

Sets the focus, gained and lost flags. If this ends up with the same widget as before focus, gained and lost flag are all set.

If no field has the focus the first one gets it.

pub fn prev(&self) -> bool

Focus the previous widget in the cycle.

Sets the focus and lost flags. If this ends up with the same widget as before it returns true and sets the focus, gained and lost flag.

If no field has the focus the first one gets it.