Struct rat_salsa::FocusFlag

pub struct FocusFlag {
    pub tag: Cell<u16>,
    pub focus: Cell<bool>,
    pub gained: Cell<bool>,
    pub lost: Cell<bool>,
}

Contains flags for the focus.

This struct is used as part of the widget state.

See HasFocusFlag, validate! and also on_gained!, on_lost!.

Fields
tag: Cell<u16>

A unique tag within one focus-cycle. It is set when the focus cycle is created. See Focus::focus

focus: Cell<bool>

Focus.

gained: Cell<bool>

This widget just gained the focus.

See on_gained!, validate!

It is reset at the beginning of all handle_xxx() calls.

lost: Cell<bool>

This widget just lost the focus.

See on_lost!, validate!

It is reset at the beginning of all handle_xxx() calls.

Implementations

impl FocusFlag

pub fn get(&self) -> bool

Has the focus.

pub fn set(&self)

Set the focus.

pub fn tag(&self) -> u16

Associated tag.

pub fn lost(&self) -> bool

Just lost the focus.

pub fn gained(&self) -> bool

Just gained the focus.