Struct rat_salsa::MouseOnly

pub struct MouseOnly;

Marker struct like DefaultKeys. This one selects an event-handler that processes only mouse events. Useful when creating your own key-bindings but not wanting to touch the mouse interactions. If this separation exists for a widget it should be called automatically by the DefaultKeys handler.