Trait rat_salsa::HandleCrossterm

pub trait HandleCrossterm<R, KeyMap = DefaultKeys> {
    // Required method
    fn handle(&mut self, event: &Event, keymap: KeyMap) -> R;
}

Handle events received by crossterm.

This one should be implemented for the state struct of a widget and can do whatever. And it can return whatever extra outcome is needed. Common usage would return a ControlUI flag as a result.

There is an extra parameter KeyMap which can be used to define more than one mapping for a widget. This can be useful when overriding the default behaviour for a widget. Two keymaps for common usage are defined in this library: DefaultKeys and MouseOnly.

    check_break!(uistate.page1.table1.handle(evt, DefaultKeys));

Remark

There is only HandleCrossterm for now, as that is what I needed. But there is no problem adding a HandleTermion, HandleTermwiz or whatever. One could add a second type parameter for this differentiation, but I think that would complicate usage unnecessarily. And any application will probably decide to use one or the other and not all of them. A widget library can easily support all of them with this scheme without some added layer of indirection and use a feature flag to select between them.

Required Methods

fn handle(&mut self, event: &Event, keymap: KeyMap) -> R