Trait rat_salsa::Input

pub trait Input<R> {
    type Request;

    // Required method
    fn perform(&mut self, req: Self::Request) -> R;
}

Execute an abstract input-action defined by a widget.

A widget can define a set of standard actions that manipulate its state. The event-handler translates the concrete events to these actions and executes them calling Input::perform. This requires that both Input and HandleCrossterm are implemented for the widget-state. As it’s not obligatory to use this extra layer this is not enforced at type-level.

Required Associated Types

type Request

Type of action request for a widget.

Required Methods

fn perform(&mut self, req: Self::Request) -> R

Perform the given action.