Trait rat_salsa::TuiApp

pub trait TuiApp {
    type Data;
    type State;
    type Action;
    type Error;

    // Required methods
    fn init(
        &self,
        data: &mut Self::Data,
        uistate: &mut Self::State,
        send: &Sender<Self::Action>
    ) -> Result<(), Self::Error>;
    fn repaint(
        &self,
        event: RepaintEvent,
        frame: &mut Frame<'_>,
        data: &mut Self::Data,
        uistate: &mut Self::State
    ) -> ControlUI<Self::Action, Self::Error>;
    fn handle_timer(
        &self,
        event: Timed,
        data: &mut Self::Data,
        uistate: &mut Self::State
    ) -> ControlUI<Self::Action, Self::Error>;
    fn handle_event(
        &self,
        event: Event,
        data: &mut Self::Data,
        uistate: &mut Self::State
    ) -> ControlUI<Self::Action, Self::Error>;
    fn run_action(
        &self,
        action: Self::Action,
        data: &mut Self::Data,
        uistate: &mut Self::State,
        send: &Sender<Self::Action>
    ) -> ControlUI<Self::Action, Self::Error>;
    fn run_task(
        &self,
        task: Self::Action,
        send: &Sender<ControlUI<Self::Action, Self::Error>>
    ) -> ControlUI<Self::Action, Self::Error>;
    fn report_error(
        &self,
        error: Self::Error,
        data: &mut Self::Data,
        uistate: &mut Self::State
    ) -> ControlUI<Self::Action, Self::Error>;

    // Provided methods
    fn get_repaint<'b>(&self, uistate: &'b Self::State) -> Option<&'b Repaint> { ... }
    fn get_timers<'b>(&self, uistate: &'b Self::State) -> Option<&'b Timers> { ... }
}

Describes the requisites of a TuiApp.

It holds no state of its own, just a way to collect the types and operations.

Required Associated Types

type Data

Application data.

type State

UI state.

type Action

Action type.

type Error

Error type.

Required Methods

fn init( &self, data: &mut Self::Data, uistate: &mut Self::State, send: &Sender<Self::Action> ) -> Result<(), Self::Error>

Do some init immediately before the event-loop starts.

fn repaint( &self, event: RepaintEvent, frame: &mut Frame<'_>, data: &mut Self::Data, uistate: &mut Self::State ) -> ControlUI<Self::Action, Self::Error>

Repaint the ui.

fn handle_timer( &self, event: Timed, data: &mut Self::Data, uistate: &mut Self::State ) -> ControlUI<Self::Action, Self::Error>

Handle a timer.

fn handle_event( &self, event: Event, data: &mut Self::Data, uistate: &mut Self::State ) -> ControlUI<Self::Action, Self::Error>

Handle an event.

fn run_action( &self, action: Self::Action, data: &mut Self::Data, uistate: &mut Self::State, send: &Sender<Self::Action> ) -> ControlUI<Self::Action, Self::Error>

Run an action.

fn run_task( &self, task: Self::Action, send: &Sender<ControlUI<Self::Action, Self::Error>> ) -> ControlUI<Self::Action, Self::Error>

Called by the worker thread to run a Task.

fn report_error( &self, error: Self::Error, data: &mut Self::Data, uistate: &mut Self::State ) -> ControlUI<Self::Action, Self::Error>

Error reporting.

Provided Methods

fn get_repaint<'b>(&self, uistate: &'b Self::State) -> Option<&'b Repaint>

Get the repaint state for this uistate.

fn get_timers<'b>(&self, uistate: &'b Self::State) -> Option<&'b Timers>

Get the timer state for this uistate.