Struct rat_salsa::widget::date_input::DateInputState

pub struct DateInputState {
    pub input: MaskedInputExtState,
    pub pattern: String,
    pub locale: Locale,
    pub non_exhaustive: NonExhaustive,
}

Fields
input: MaskedInputExtStatepattern: Stringlocale: Localenon_exhaustive: NonExhaustive
Implementations

impl DateInputState

pub fn reset(&mut self)

Reset to empty.

pub fn set_format<S: AsRef<str>>(&mut self, pattern: S) -> Result<(), Error>

chrono format string.

generates a mask according to the format and overwrites whatever set_mask() did.

pub fn set_formats<S: AsRef<str>>( &mut self, pattern: S, locale: Locale ) -> Result<(), Error>

chrono format string.

generates a mask according to the format and overwrites whatever set_mask() did.

pub fn set_display_mask<S: Into<String>>(&mut self, s: S)

Overlay for empty fields.

pub fn set_mask<S: AsRef<str>>(&mut self, s: S) -> Result<(), Error>

Mask as defined by [MaskedInput]

pub fn value(&self) -> Result<NaiveDate, ParseError>

pub fn set_value(&mut self, date: NaiveDate)

pub fn select_all(&mut self)