Module rat_salsa::widget::mask_input

Text input with an input mask.

Can do the usual insert/delete/move operations.

Text selection

Scrolls with the cursor.

Can set the cursor or use its own block cursor.

Can show an indicator for invalid input.

Accepts an input mask:

0: can enter digit, display as 0

9: can enter digit, display as space

#: digit, plus or minus sign, display as space

-: sign

+: sign, positive is ‘+’, negative is ‘-’, not localized.

. and ,: decimal and grouping separators

H: must enter a hex digit, display as 0

h: can enter a hex digit, display as space

O: must enter an octal digit, display as 0

o: can enter an octal digit, display as space

D: must enter a decimal digit, display as 0

d: can enter a decimal digit, display as space

l: can enter letter, display as space

a: can enter letter or digit, display as space

c: can enter character or space, display as space

_: anything, display as space

<space>, :, ;, /: separator characters move the cursor when entered.

\: escapes the following character and uses it as a separator.

all other ascii characters a reserved.

other unicode characters can be used as separators without escaping.

Accepts a display overlay used instead of the default chars of the input mask.

use rat_salsa::widget::mask_input::MaskedInputState;

let mut date_state = MaskedInputState::new();
date_state.set_mask("99/99/9999");
date_state.set_display_mask("mm/dd/yyyy");

let mut creditcard_state = MaskedInputState::new();
creditcard_state.set_mask("dddd dddd dddd dddd");

Modules

core

Structs

MaskedInput

Text input widget with input mask.

MaskedInputState

State of the input-mask.

MaskedInputStyle

Combined style.

Enums

InputRequest

Mapping from events to abstract editing requests.