Struct rat_salsa::widget::mask_input::core::InputMaskCore

pub struct InputMaskCore { /* private fields */ }

Text editing core.

Implementations

impl InputMaskCore

pub fn new() -> Self

pub fn new_with_symbols(sym: &Rc<NumberSymbols>) -> Self

pub fn tokens(&self) -> &[MaskToken]

Tokens used for the mask.

pub fn reset(&mut self)

Reset value but not the mask and width. Resets offset and cursor position too.

pub fn offset(&self) -> usize

Offset

pub fn set_offset(&mut self, offset: usize)

Change the offset.

Ensures the cursor is visible and modifies any given offset. Ensures the offset is not beyond the length.

pub fn width(&self) -> usize

Display width

pub fn set_width(&mut self, width: usize)

Display width

pub fn cursor(&self) -> usize

Cursor position as grapheme-idx.

pub fn anchor(&self) -> usize

pub fn has_selection(&self) -> bool

pub fn is_select_all(&self) -> bool

pub fn selection(&self) -> Range<usize>

pub fn set_cursor(&mut self, cursor: usize, extend_selection: bool)

pub fn set_default_cursor(&mut self)

Place cursor at decimal separator, if any. 0 otherwise.

pub fn set_num_symbols(&mut self, sym: &Rc<NumberSymbols>)

Set the decimal separator and other symbols. Only used for rendering and to map user input. The value itself uses “.”

pub fn set_mask<S: AsRef<str>>(&mut self, s: S) -> Result<(), Error>

Changes the mask. Resets the value to a default.

pub fn mask(&self) -> String

Return the mask.

pub fn debug_mask(&self) -> String

Return the mask.

pub fn set_display_mask<S: Into<String>>(&mut self, s: S)

Set the mask that is shown.

pub fn display_mask(&self) -> String

Display mask

pub fn set_value<S: Into<String>>(&mut self, s: S)

Sets the value. No checks if the value conforms to the mask. If the value is too short it will be filled with space. if the value is too long it will be truncated.

pub fn value(&self) -> &str

Value

pub fn compact_value(&self) -> String

Value without whitespace and grouping separators. Might be easier to parse.

pub fn is_empty(&self) -> bool

No value different from the default.

pub fn len(&self) -> usize

Length

pub fn rendered(&self) -> &str

Rendered string for display.

pub fn render_value(&mut self)

Create the rendered value.

pub fn render_condensed_value(&mut self)

Create the rendered value.

pub fn next_word_boundary(&self) -> usize

Next boundary.

pub fn prev_word_boundary(&self) -> usize

Previous boundary.

pub fn advance_cursor(&mut self, c: char)

Start at the cursor position and find a valid insert position for the input c. Put the cursor at that position.

pub fn insert_char(&mut self, c: char) -> Result<(), Error>

Insert the char if it matches the cursor mask and the current section is not full.

advance_cursor() must be called before for correct functionality.

Otherwise: your mileage might vary.

pub fn remove_selection(&mut self, selection: Range<usize>) -> Result<(), Error>

Remove the selection.

pub fn remove_prev(&mut self) -> Result<(), Error>

Remove the previous char.

pub fn remove_next(&mut self) -> Result<(), Error>

Remove the next char.